created dedicated modules fix, colors for specific functions renamed publib.py as main.py
