+++ publib/__init__.py
@@ -5,4 +5,5 @@ publib
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from .publib import set_style, fix_style
+from .main import set_style, fix_style
+from .tools.colors import colors
