+++ publib/test/test_functions.py
@@ -8,7 +8,8 @@ Created on Sun Apr  8 18:26:37 2018
 from __future__ import absolute_import, print_function
 
 from publib import set_style, fix_style
-from publib.tools import reset, fix_bold_TimesNewRoman
+from publib.tools.tools import reset
+from publib.tools.fix import fix_bold_TimesNewRoman
 import matplotlib as mpl
 
 # %% Test routines
