+++ publib/tools/colors.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+"""
+
+"""
+
+
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+import matplotlib as mpl
+
+# Colors available for import
+colors = ['#5DA5DA',
+          '#FAA43A',
+          '#60BD68',
+          '#F17CB0',
+          '#B2912F',
+          '#B276B2',
+          '#DECF3F',
+          '#F15854',
+          '#4D4D4D']
